backend/requirements.txt
# Note: Django 5+ requires Python 3.10+. Using 4.2 LTS for Python 3.9.
# When Python 3.12 is installed, upgrade to: django>=5.0, djangorestframework>=3.15
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
pgvector>=0.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
